+++ src/git.js
@@ -1,59 +1,61 @@
 const callLimiter = require("./call-limiter.js");
-const { add, git, update } = require("./constants.js");
+const { add, git, types, update } = require("./constants.js");
 
 const BUFFER_DELAY = 150;
-const stageBuffers = {};
-const defaultOptions = { env: process.env };
+const DEFAULT_OPTIONS = { env: process.env };
+const STAGE_BUFFERS = {};
+
 const limiter = callLimiter.new({ concurrency: 1 });
 
+let _available = null;
+
 /**
- * Execute arbitrary git commands one at the time. When
- * multiple calls are performed in quick succession the
- * git commands are automatically rate limited.
+ * Execute arbitrary git commands one at the time. When multiple calls are
+ * made in quick succession the commands are automatically rate limited.
  *
- * Example:
- * ```
+ * @private
+ * @param  {Array} args - The arguments for the command in order.
+ * @param  {String} gitCwd - Relative path to the root of the git repository.
+ * @param  {Function} callback - Function called on completion.
+ * @example
+ * const options = {gitCwd: "../"}
  * gitExecute(['add', 'myfile.txt'], options, callback);
  * // -> executes: $ git add myfile.txt
- * ```
- *
- * @param  {Array}    args     The arguments for the command in order.
- * @param  {Object}   _options Options for the command.
- *                      - gitCwd: directory containing the '.git' folder.
- * @param  {Function} callback Function called on completion.
  */
 function gitExecute(args, _options, callback) {
   const exec = require("child_process").execFile;
 
-  if (!_export.available) {
+  if (!isGitAvailable()) {
     throw new Error("git not found on your system.");
   }
 
-  let options = Object.assign(defaultOptions, {
-    cwd: _options.gitCwd || __dirname,
+  const options = Object.assign(DEFAULT_OPTIONS, {
+    cwd: _options.gitCwd,
   });
 
   return limiter.schedule(exec, git, args, options, callback);
 }
 
 /**
- * Get a buffer for a particular stream (based on the
- * id) to which additional files can be pushed such
- * that in the end only a single `git add` command is
- * executed.
+ * Get a buffer for a particular stream (based on the id) to which files can be
+ * pushed such that in the end only a single `git add` command is executed.
  *
- * @param {Any} id        A unique identifier for a stream.
- * @return {StageBuffer}  The buffer that can be used to add files.
+ * @private
+ * @param {Any} id - A unique identifier for a stream.
+ * @return {StageBuffer} The buffer that can be used to add files.
+ * @example
+ * const stageBuffer = getStageBufferForStream(42);
+ * stageBuffer.push(file, config, callback)
  */
 function getStageBufferForStream(id) {
   const debounce = require("debounce");
 
-  if (stageBuffers[id] === undefined) {
+  if (STAGE_BUFFERS[id] === undefined) {
     const callbacks = [];
     const files = [];
 
     const db = debounce(function(config) {
-      stageBuffers[id] = undefined;
+      STAGE_BUFFERS[id] = undefined;
 
       const args = [add];
       if (config.stagedOnly) args.push(update);
@@ -66,7 +68,7 @@ function getStageBufferForStream(id) {
       });
     }, BUFFER_DELAY);
 
-    stageBuffers[id] = {
+    STAGE_BUFFERS[id] = {
       push: function(file, config, callback) {
         callbacks.push(callback);
         files.push(file);
@@ -75,42 +77,59 @@ function getStageBufferForStream(id) {
     };
   }
 
-  return stageBuffers[id];
+  return STAGE_BUFFERS[id];
 }
 
-/* === EXPORT === */
+/**
+ * Check whether or not git is available on the system.
+ *
+ * @private
+ * @return {Boolean} True if git is available on the system, false otherwise.
+ */
+function isGitAvailable() {
+  if (_available === null) {
+    const which = require("which");
 
-const _export = {};
+    const result = which.sync(git, { nothrow: true });
+    _available = !(result === null);
+  }
 
-let _available = null;
-Object.defineProperty(_export, "available", {
-  get: function() {
-    if (_available === null) {
-      const which = require("which");
-
-      let result = which.sync(git, { nothrow: true });
-      _available = !(result === null);
-    }
-
-    return _available;
-  },
-});
+  return _available;
+}
 
 /**
- * @param  {String}   file     The path to the file to stage.
- * @param  {Object}   config   Configuration of the stage action.
- *                      - gitCwd: directory containing the '.git' folder.
- *                      - stagedOnly: only stage previously staged files.
- * @param  {Any}      streamId A unique identifier for a stream.
- * @param  {Function} callback The function to call on completion.
+ * Provides an API to interact with git.
+ * @module git
  */
-_export.stage = function(file, config, streamId, callback) {
-  if (typeof file !== "string") {
-    callback("file must be a string.");
-  } else {
-    const buffer = getStageBufferForStream(streamId);
-    buffer.push(file, config, callback);
-  }
+module.exports = {
+  /**
+   * Stage a file in a git repository.
+   *
+   * @param  {String} file - Path to the file to stage.
+   * @param  {String} gitCwd - Relative path to the root of the git repository.
+   * @param  {Boolean} stagedOnly - Only stage previously staged files.
+   * @param  {Any} streamId - An identifier to bundle multiple files.
+   * @param  {Function} callback - The function to call on completion.
+   * @throws {TypeError} Callback is not a function.
+   * @throws {Error} The git command is not present.
+   * @example
+   * const config = {gitCwd: "../", stagedOnly: false};
+   * stage('my-file.js', config, 42, error => {
+   *  if (error) {
+   *    console.log('Failed.');
+   *  } else {
+   *    console.log('Success!');
+   *  }
+   * });
+   */
+  stage: function(file, config, streamId, callback) {
+    if (typeof file !== types.string) {
+      callback("file must be a string.");
+    } else if (typeof callback !== types.function) {
+      throw new TypeError("callback is not a function");
+    } else {
+      const stageBuffer = getStageBufferForStream(streamId);
+      stageBuffer.push(file, config, callback);
+    }
+  },
 };
-
-module.exports = _export;